Moves topology handling to provisioner
Also removes cfg testing since we can now work without cfg file
