+++ cli/spec.py
@@ -70,7 +70,6 @@ def parse_args(app_settings_dir, args=None):
     sub_parser_options = subparsers_options.get(clg_args['command0'], {})
 
     override_default_values(clg_args, sub_parser_options)
-    process_topology_args(clg_args, app_settings_dir)
     return clg_args
 
 
@@ -112,36 +111,6 @@ def override_default_values(clg_args, sub_parser_options):
             conflict_resolver=utils.ConflictResolver.none_resolver)
 
         _check_required_arguments(sub_parser_options, clg_args)
-
-
-def process_topology_args(clg_args, app_settings_dir):
-    """
-    Merges topology files in a single topology dict.
-
-    :param clg_args: Dictionary based on cmd-line args parsed by clg
-    :param app_settings_dir: path to the base directory holding the
-        application's settings. App can be provisioner\installer\tester
-        and the path would be: settings/<app_name>/
-    """
-
-    # post process topology
-    if clg_args.get("topology") is not None:
-        topology_dict = {}
-        for topology_item in clg_args['topology'].split(','):
-            if '_' in topology_item:
-                number, node_type = topology_item.split('_')
-            else:
-                raise exceptions.IRConfigurationException(
-                    "Topology node should be in format  <number>_<node role>. "
-                    "Current value: '{}' ".format(topology_item))
-            # todo(obaraov): consider moving topology to config on constant.
-            node_file = os.path.join(
-                app_settings_dir, 'topology', node_type + '.yaml')
-            with open(node_file) as stream:
-                topology_dict[node_type] = yaml.load(stream)
-                topology_dict[node_type]['amount'] = int(number)
-
-        clg_args['topology'] = topology_dict
 
 
 def _check_required_arguments(command_args, clg_args):
