--- tests/test_cwd/infrared.cfg
@@ -1,2 +1,0 @@
-[defaults]
-settings  = settings
